+++ ryoku/shell/quickshell/shell/modules/bar/barstyles/qsbar/Scene.qml
@@ -11,6 +11,7 @@
 
 import Quickshell
 import QtQuick
+import shell.services
 import "../../../../services/lib/screens.js" as Screens
 import "."
 import "core"
@@ -31,6 +32,7 @@ Item {
     // the bar system is never hosted twice.
     readonly property bool isPrimary: {
         var list = Screens.uniqueByName(Quickshell.screens)
+            .filter(screen => Config.barStyleFor(screen.name) === "qsbar")
         return list.length > 0 && !!sceneRoot.modelData
             && list[0].name === sceneRoot.modelData.name
     }